Refactor: libcrmcommon: separate attrd client utilities into own source file
